sim/requirements.txt
paho-mqtt
numpy
airsim
msgpack-rpc-python
opencv-python
Pillow
